requirements.txt
# Django and DRF - Python 3.11 compatible
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3
gunicorn==21.2.0

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# File handling - Stable versions for Python 3.11
Pillow==11.3.0
pandas==2.3.1
openpyxl==3.1.2
xlrd==2.0.1

# OCR and Image Processing - Skip problematic packages for now
pytesseract==0.3.10

# Task Queue (Django-Q)
django-q==1.3.9
redis==3.5.3

# Email
django-ses==3.5.0

# Environment
python-decouple==3.8
python-dotenv==1.0.0

# API Documentation
drf-spectacular==0.26.5

# Utilities
requests==2.31.0
httpx==0.25.2

# Monitoring and Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

# Storage (for production)
django-storages==1.14.2

# Static files for production
whitenoise==6.6.0

# Security
django-ratelimit==4.1.0
django-axes==6.1.1

# Barcode generation
python-barcode==0.15.1
qrcode==7.4.2

# Data validation
cerberus==1.3.5
jsonschema==4.19.2

# Testing - Skip for deployment
# pytest==7.4.3
# pytest-django==4.7.0
# factory-boy==3.3.0
# faker==20.1.0

# Development - Skip for deployment
# black==23.10.1
# flake8==6.1.0
# isort==5.12.0

# POS Integration
squareup>=33.0.0,<34.0.0

# Caching
django-redis==5.4.0

# Timezone handling
pytz==2023.3

# Math and calculations - Skip numpy/scipy for now as they often cause issues
# numpy==1.24.4
# scipy==1.10.1

# Export formats
reportlab==4.0.7
xlsxwriter==3.1.9

# Encryption
cryptography==42.0.8